pyproject.toml
[project]
name = "trainingload"
version = "1.0.0"
description = "TrainingLoad - Evidence-based training load monitoring system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.1.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.0",
    "requests>=2.32.5",
    "apscheduler>=3.10.4",
    "python-multipart>=0.0.6",
    "numpy>=2.3.2",
    "psutil>=5.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
